platform changed to MSVC2019(x64) + freeglut + fltk-1.3.5 + opencv-4.2.0

+++ GUI/GraphWindowCP.cpp
@@ -494,13 +494,13 @@ int GraphWindowCP::handle(int event)
 					flg_psize = 7;
 				} else if( cridx>-1 && cpidx>-1 ){
 					flg_psize = 6;
-				} else if( abs( (ox+psx*wscl) - push_x ) < thres ){
+				} else if( abs( (int)(ox+psx*wscl) - push_x ) < thres ){
 					flg_psize = 1;
-				} else if( abs( (ox+pex*wscl) - push_x ) < thres ){
+				} else if( abs( (int)(ox+pex*wscl) - push_x ) < thres ){
 					flg_psize = 2;
-				} else if( abs( (oy+psy*hscl) - push_y ) < thres ){
+				} else if( abs( (int)(oy+psy*hscl) - push_y ) < thres ){
 					flg_psize = 3;
-				} else if( abs( (oy+pey*hscl) - push_y ) < thres ){
+				} else if( abs( (int)(oy+pey*hscl) - push_y ) < thres ){
 					flg_psize = 4;
 				} else {
 					flg_psize = 0;
